requirements.txt
accelerate==1.7.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
et_xmlfile==2.0.0
filelock==3.13.1
Flask==3.1.1
fsspec==2024.6.1
huggingface-hub==0.33.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.5.1
lxml==5.4.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.3.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pgvector==0.4.1
pillow==11.2.1
psutil==7.0.0
psycopg2==2.9.10
PyMuPDF==1.26.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==4.1.0
six==1.17.0
soupsieve==2.7
sympy==1.13.3
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.1+cu118
tqdm==4.67.1
transformers==4.52.4
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
XlsxWriter==3.2.3
